use int

+++ mmpose/datasets/datasets/body3d/body3d_mview_direct_panoptic_dataset.py
@@ -211,7 +211,7 @@ class Body3DMviewDirectPanopticDataset(Kpt3dMviewRgbImgDirectDataset):
                             dtype=np.float32)
 
                         cnt = 0
-                        person_ids = -np.ones(self.max_persons, dtype=np.int32)
+                        person_ids = -np.ones(self.max_persons, dtype=int)
                         for body in bodies:
                             if cnt >= self.max_persons:
                                 break
